requirements.txt
flask==2.2.5
gunicorn==21.2.0
meross_iot==0.4.6.2
python-dotenv==0.21.1
requests==2.31.0